+++ modules/special_functions/tests/test_factor.m
@@ -23,35 +23,23 @@
 % License along with this program. If not, see <http://www.gnu.org/licenses/>.
 % LICENCE_BLOCK_END
 %=============================================================================
-function p = primes(n)
-  narginchk(1, 1)
-  nargoutchk(0, 1)
-  
-  if ~isscalar(n)
-    error(_('Wrong size for argument #1: scalar expected.'));
-  end
-  if ~isfinite(n)
-    error(_('Wrong value for argument #1: finite value expected.'));
-  end
-  if ~isreal(n)
-    error(_('Wrong value for argument #1: real value expected.'));
-  end
-  
-  if n < 2
-    p = zeros(1, 0, class(n));
-    return
-  end
-  N = floor(double(n));
-  p = [1:2:N];
-  q = length(p);
-  p(1) = 2;
-  for k = 3:2:sqrt(N)
-    if p((k+1)/2)
-      p([((k*k+1)/2):k:q]) = 0;
-    end
-  end
-  p = p(find(p >= 0));
-  p = p(find(p ~= 0));
-  p = cast(p, class(n));
-end
+assert_isequal(nargin('factor'), 1);
+assert_isequal(nargout('factor'), 1);
+%=============================================================================
+R = factor(200);
+REF = [2     2     2     5     5];
+assert_isequal(R, REF);
+%=============================================================================
+R = factor(single(200));
+REF = single([2     2     2     5     5]);
+assert_isequal(R, REF);
+%=============================================================================
+n = uint16(138);
+R = factor(n);
+REF = uint16([ 2    3   23]);
+assert_isequal(R, REF);
+%=============================================================================
+R = factor(3);
+REF = 3;
+assert_isequal(R, REF);
 %=============================================================================